Replace PageWrapper HOC with Layout files
