+++ apps/web/app/event-types-1/page.tsx
@@ -1,11 +1,7 @@
-import EventTypesPage from "@pages/event-types";
+import EventTypes from "@pages/event-types";
 import type { Metadata } from "next";
-import { headers } from "next/headers";
 
-import { getLayout } from "@calcom/features/MainLayout";
 import { IS_CALCOM, WEBAPP_URL } from "@calcom/lib/constants";
-
-import PageWrapper from "@components/PageWrapperAppDir";
 
 export const metadata: Metadata = {
   viewport: "width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0",
@@ -19,19 +15,6 @@ export const metadata: Metadata = {
   },
 };
 
-export default function EventTypesPageWrapped(props: Record<string, unknown>) {
-  const h = headers();
-
-  const nonce = h.get("x-nonce") ?? undefined;
-
-  return (
-    <PageWrapper
-      getLayout={getLayout}
-      requiresLicense={false}
-      pageProps={props}
-      nonce={nonce}
-      themeBasis={null}>
-      <EventTypesPage />
-    </PageWrapper>
-  );
+export default function EventTypesPage(props: Record<string, unknown>) {
+  return <EventTypes {...props} />;
 }
